'''Windows''' the core now compiles and most tests pass!  * There is an outstanding issue with ConfigNode::fromFile method, either it doesn't work, or the test can't feed it the proper path (yes the path has been tried windows style)  * The major work here was porting the time code over to windows. This included finding a windows version of gettimeofday  * The current method is not nearly as accurate on Linux, Real Time Linux, or Mac, but it will be good enough for windows based testing  * Added missing pattern export header

+++ packages/core/test/src/QueueTest.cpp
@@ -44,7 +44,7 @@ boost::mutex io_mutex;
 void dataProducer()
 {
     unsigned char c = 'A';
-    for(uint i = 0; i < NUM_CHARS;)
+    for(size_t i = 0; i < NUM_CHARS;)
     {
         char out = 0;
         if((c >= 'A' && c <= 'Z'))
